Add controllers using mvc
