+++ routes/shop.js
@@ -5,18 +5,10 @@ const express = require('express');
 const rootDir = require('../util/path');
 const adminData = require('./admin');
 
+const productsController = require('../controllers/products');
+
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-  const products = adminData.products;
-  res.render('shop', {
-    prods: products,
-    pageTitle: 'Shop',
-    path: '/',
-    hasProducts: products.length > 0,
-    activeShop: true,
-    productCSS: true
-  });
-});
+router.get('/', productsController.getProducts);
 
 module.exports = router;
